requirements.txt
holidays
tda-api
webdriver_manager
